C : 게시글 생성 시, Tag를 추가하지 않는 경우에 대한 예외처리

+++ simproject/src/main/java/org/project/simproject/service/TagService.java
@@ -77,6 +77,9 @@ public class TagService {       // DTO 이용하도록 추후에 수정
     @Transactional
     public void deleteRelations(Article article) {
         List<ArticleTagRelation> relations = articleTagRelationRepository.findArticleTagRelationsByArticle(article);
+
+        if (relations.isEmpty()) return;
+
         articleTagRelationRepository.deleteAll(relations);
 
         for (ArticleTagRelation relation : relations) {
